requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10
django-localflavor==1.5.2
django-money==0.12
ecdsa==0.13
future==0.16.0
idna==2.6
py-moneyed==0.7.0
pycrypto==2.6.1
python-dateutil==2.6.1
python-jose==1.3.2
python-quickbooks==0.7.0
pytz==2017.2
rauth==0.7.3
requests==2.13.0
simplejson==3.11.1
six==1.11.0
urllib3==1.22
